video_uploader/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
kafka-python==2.2.2
mysqlclient==2.2.7
pillow==11.1.0
pycryptodome==3.22.0
PyJWT==2.10.1
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.1